1. modify iostream/inc/g_io_exception.h;

+++ module/gcommon/iostream/inc/g_io_exception.h
@@ -23,7 +23,7 @@
 inline std::string EXCEPTION_DESCRIPTION(const std::string& desc)
 {
 	return (std::string("File:") + std::string(__FILE__) + \
-	std::string(", Line:") + std::to_string(__LINE__) +\
+	std::string(", Line:") + std::to_string(static_cast<long long>(__LINE__)) +\
 	std::string(", exception:") + std::string(desc));
 }
